% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdis_maps.R
\name{catdis.plot.heatmap.school}
\alias{catdis.plot.heatmap.school}
\title{TBD}
\usage{
catdis.plot.heatmap.school(
  base_map = map.atlantic(),
  catdis_data,
  school_type,
  num_breaks = 5
)
}
\arguments{
\item{base_map}{TBD}

\item{catdis_data}{TBD}

\item{school_type}{TBD}

\item{num_breaks}{TBD}
}
\value{
TBD
}
\description{
TBD
}
