% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdis_maps.R
\name{catdis.plot.piemap}
\alias{catdis.plot.piemap}
\title{TBD}
\usage{
catdis.plot.piemap(
  base_map = map.atlantic(projection = iccat.pub.maps::CRS_EQUIDISTANT),
  catdis_data,
  gears_to_keep = NULL,
  default_radius = pi,
  max_catch = NA,
  center_pies = TRUE,
  legend.x = -90,
  legend.y = -25,
  crs = iccat.pub.maps::CRS_EQUIDISTANT
)
}
\arguments{
\item{base_map}{TBD}

\item{catdis_data}{TBD}

\item{gears_to_keep}{TBD}

\item{default_radius}{TBD}

\item{max_catch}{TBD}

\item{center_pies}{TBD}

\item{legend.x}{TBD}

\item{legend.y}{TBD}

\item{crs}{TBD}
}
\value{
TBD
}
\description{
TBD
}
