% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdis_maps.R
\name{catdis.plot.piemap.school}
\alias{catdis.plot.piemap.school}
\title{TBD}
\usage{
catdis.plot.piemap.school(
  base_map = map.atlantic(),
  catdis_data,
  default_radius = pi,
  max_catch = NA,
  center_pies = TRUE,
  legend.x = -90,
  legend.y = -25
)
}
\arguments{
\item{base_map}{TBD}

\item{catdis_data}{TBD}

\item{default_radius}{TBD}

\item{max_catch}{TBD}

\item{center_pies}{TBD}

\item{legend.x}{TBD}

\item{legend.y}{TBD}
}
\value{
TBD
}
\description{
TBD
}
