% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_area.R
\name{t1nc.plot.area_sampling_areas}
\alias{t1nc.plot.area_sampling_areas}
\title{TBD}
\usage{
t1nc.plot.area_sampling_areas(t1nc_data, relative = FALSE)
}
\arguments{
\item{t1nc_data}{TBD}

\item{relative}{TBD}
}
\value{
TBD
}
\description{
TBD
}
