% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_bar.R
\name{t1nc.plot.bar_species_gear_groups}
\alias{t1nc.plot.bar_species_gear_groups}
\title{Default plot function to display annual time series of T1 nominal catch data as a bar chart categorized by species' gear group
(see the \code{GearGrpBySpecies} and \code{Species} table in \code{\link{iccat.dev.base::DATABASE_T1}})}
\usage{
t1nc.plot.bar_species_gear_groups(
  t1nc_data,
  max_categories = NA,
  relative = FALSE
)
}
\arguments{
\item{t1nc_data}{the T1 nominal catch data as retrieved through \code{\link{iccat.dev.data::t1nc}}}

\item{max_categories}{the maximum number of categories to display. Everything else will be placed under the _Other_ category}

\item{relative}{whether or not the bars should represent the relative composition of annual catches, or their absolute values}
}
\value{
a bar chart of the provided data with the given configuration
}
\description{
Default plot function to display annual time series of T1 nominal catch data as a bar chart categorized by species' gear group
(see the \code{GearGrpBySpecies} and \code{Species} table in \code{\link{iccat.dev.base::DATABASE_T1}})
}
