% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_bar.R
\name{t1nc.plot.bar_species_gears}
\alias{t1nc.plot.bar_species_gears}
\title{TBD}
\usage{
t1nc.plot.bar_species_gears(t1nc_data, max_categories = NA, relative = FALSE)
}
\arguments{
\item{t1nc_data}{TBD}

\item{max_categories}{TBD}

\item{relative}{TBD}
}
\value{
TBD
}
\description{
TBD
}
