% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_bar.R
\name{t1nc.plot.bar_stocks}
\alias{t1nc.plot.bar_stocks}
\title{Default plot function to display annual time series of T1 nominal catch data as a bar chart categorized by species' stocks}
\usage{
t1nc.plot.bar_stocks(t1nc_data, relative = FALSE)
}
\arguments{
\item{t1nc_data}{the T1 nominal catch data as retrieved through \code{\link{iccat.dev.data::t1nc}}}

\item{relative}{whether or not the bars should represent the relative composition of annual catches, or their absolute values}
}
\value{
a bar chart of the provided data with the given configuration
}
\description{
Default plot function to display annual time series of T1 nominal catch data as a bar chart categorized by species' stocks
}
