% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_line.R
\name{t1nc.plot.line_gears}
\alias{t1nc.plot.line_gears}
\title{TBD}
\usage{
t1nc.plot.line_gears(t1nc_data, max_categories = NA)
}
\arguments{
\item{t1nc_data}{TBD}

\item{max_categories}{TBD}
}
\value{
TBD
}
\description{
TBD
}
