% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_line.R
\name{t1nc.plot.line_species_gear_groups}
\alias{t1nc.plot.line_species_gear_groups}
\title{TBD}
\usage{
t1nc.plot.line_species_gear_groups(
  t1nc_data,
  max_categories = NA,
  relative = FALSE
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{max_categories}{TBD}

\item{relative}{TBD}
}
\value{
TBD
}
\description{
TBD
}
