% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_pareto.R
\name{t1nc.plot.pareto}
\alias{t1nc.plot.pareto}
\title{TBD}
\usage{
t1nc.plot.pareto(
  t1nc_data,
  x_column,
  x_name,
  category_column,
  category_name,
  category_levels,
  category_colors,
  max_x,
  max_categories,
  vertical
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{x_column}{TBD}

\item{x_name}{TBD}

\item{category_column}{TBD}

\item{category_name}{TBD}

\item{category_levels}{TBD}

\item{category_colors}{TBD}

\item{max_x}{TBD}

\item{max_categories}{TBD}

\item{vertical}{TBD}
}
\value{
TBD
}
\description{
TBD
}
