% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_pareto.R
\name{t1nc.plot.pareto_fleet_gears}
\alias{t1nc.plot.pareto_fleet_gears}
\title{Default Pareto charts of T1 nominal catch data, with fleets as X-axis categories, and gear groups as catch component categories.}
\usage{
t1nc.plot.pareto_fleet_gears(
  t1nc_data,
  max_x = 10,
  max_categories = 10,
  vertical = FALSE
)
}
\arguments{
\item{t1nc_data}{the T1 nominal catch data as retrieved through \code{\link{iccat.dev.data::t1nc}}}

\item{max_x}{the maximum number of categories being displayed on the X axis. Everything else will go under the _All others_ label}

\item{max_categories}{the maximum number of catch categories. Everything else will go under the _All others_ label}

\item{vertical}{to display of the chart vertically}
}
\value{
a Pareto chart of T1 nominal catches configured according to the provided parameters
}
\description{
Default Pareto charts of T1 nominal catch data, with fleets as X-axis categories, and gear groups as catch component categories.
}
