% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_area.R
\name{t1nc.plot.area}
\alias{t1nc.plot.area}
\title{Generic plots of annual time series T1 nominal catch data as area charts (either absolute or relative) categorized
through one of the T1 nominal catch attributes.}
\usage{
t1nc.plot.area(
  t1nc_data,
  relative = FALSE,
  category_column,
  ref_categories = NULL,
  max_categories = NA,
  other_category_label = NA,
  legend_title,
  x_breaks_every = 10,
  colors
)
}
\arguments{
\item{t1nc_data}{the T1 nominal catch data as retrieved through \code{\link{iccat.dev.data::t1nc}}}

\item{relative}{whether or not the areas should represent the relative composition of annual catches, or their absolute values}

\item{category_column}{the category column among those included in the T1 nominal catch dataset}

\item{ref_categories}{an optional list of reference category values, If not provided, it is calculated from the actual data}

\item{max_categories}{the maximum number of categories to display. Everything else will be placed under the _Other_ category
'@param other_category_label the _Other_ category label
'@param legend_title the legend title}

\item{x_breaks_every}{the number of years at which X axis breaks should be placed}

\item{colors}{a color table for the selected category, i.e., a table with three columns: \code{_categoryColumn_CODE}, \code{FILL}, \code{OUTLINE}, with the latter two being the RGB colors for the fill and outline components of each bar}
}
\value{
an area chart of the provided data with the given configuration
}
\description{
Generic plots of annual time series T1 nominal catch data as area charts (either absolute or relative) categorized
through one of the T1 nominal catch attributes.
}
