% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_area.R
\name{t1nc.plot.area_catch_types}
\alias{t1nc.plot.area_catch_types}
\title{Default plot function to display annual time series of T1 nominal catch data as an area chart categorized by catch type (retained / landed / discarded dead / etc.)}
\usage{
t1nc.plot.area_catch_types(t1nc_data, relative = FALSE)
}
\arguments{
\item{t1nc_data}{the T1 nominal catch data as retrieved through \code{\link{iccat.dev.data::t1nc}}}

\item{relative}{whether or not the areas should represent the relative composition of annual catches, or their absolute values}
}
\value{
an area chart of the provided data with the given configuration
}
\description{
Default plot function to display annual time series of T1 nominal catch data as an area chart categorized by catch type (retained / landed / discarded dead / etc.)
}
