% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_area.R
\name{t1nc.plot.area_sampling_areas}
\alias{t1nc.plot.area_sampling_areas}
\title{Default plot function to display annual time series of T1 nominal catch data as an area chart categorized by species' sampling areas}
\usage{
t1nc.plot.area_sampling_areas(t1nc_data, relative = FALSE, max_categories = 16)
}
\arguments{
\item{t1nc_data}{the T1 nominal catch data as retrieved through \code{\link{iccat.dev.data::t1nc}}}

\item{relative}{whether or not the areas should represent the relative composition of annual catches, or their absolute values}

\item{max_categories}{the maximum number of categories to display. Everything else will be placed under the _Other_ category}
}
\value{
an area chart of the provided data with the given configuration
}
\description{
Default plot function to display annual time series of T1 nominal catch data as an area chart categorized by species' sampling areas
}
