% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_line.R
\name{t1nc.plot.line}
\alias{t1nc.plot.line}
\title{TBD}
\usage{
t1nc.plot.line(
  t1nc_data,
  category_column = "GearGrp",
  x_breaks_every = 10,
  colors
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{category_column}{TBD}

\item{x_breaks_every}{TBD}

\item{colors}{TBD}
}
\value{
TBD
}
\description{
TBD
}
