% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_line.R
\name{t1nc.plot.line_sampling_areas}
\alias{t1nc.plot.line_sampling_areas}
\title{TBD}
\usage{
t1nc.plot.line_sampling_areas(t1nc_data)
}
\arguments{
\item{t1nc_data}{TBD}
}
\value{
TBD
}
\description{
TBD
}
