% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc_plots_pareto.R
\name{t1nc.plot.pareto_fleet_gears}
\alias{t1nc.plot.pareto_fleet_gears}
\title{TBD}
\usage{
t1nc.plot.pareto_fleet_gears(
  t1nc_data,
  max_x = 10,
  max_categories = 10,
  vertical = FALSE
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{max_x}{TBD}

\item{max_categories}{TBD}

\item{vertical}{TBD}
}
\value{
TBD
}
\description{
TBD
}
